--- facebook/games/__init__.py
@@ -1,2 +1,0 @@
-""" This package contains all game related stuff. """
-
