--- facebook/games/models.py
@@ -1,5 +1,0 @@
-#coding: utf-8
-
-
-
-
